requirements.txt
Django==1.11.3
djangorestframework==3.8.2
Markdown==3.0.1
django-filter==2.0.0
mysqlclient==1.3.14
Pillow==5.3.0

#xadmin https://github.com/sshwsfc/xadmin
django-crispy-forms==1.7.2
django-import-export==1.1.0
django-reversion==3.0.2
django-formtools==2.1
future==0.17.1
httplib2==0.12.0
six==1.11.0
xlwt==1.3.0
XlsxWriter==1.1.2

# drf
coreapi==2.3.3
django-guardian==1.4.9

django-cors-headers==2.4.0 # 跨域设置  配置详见github

djangorestframework-jwt==1.11.0  # django-rest-framework-jwt http://getblimp.github.io/django-rest-framework-jwt/
pycryptodome==3.7.2    # 安装加密需要的工具包

# celery
django-celery==3.2.2
redis==2.10.6
flower==0.9.2

# drf Cache 缓存
drf-extensions==0.3.1

# 配置redis缓存
django-redis==4.10.0

# 配置django用户环境变量
django-dotenv==1.4.2

# 集成sentry日志收集服务器
raven==6.10.0


